apps/cli/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pqcbench-cli"
version = "0.1.0"
description = "CLI for running PQC benchmarks and ACVP validation"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Investigation Team"}]
dependencies = [
  "pqcbench-core>=0.1.0",
  "pqcbench-rsa>=0.1.0",
  "pqcbench-liboqs>=0.1.0",
  "typer>=0.12.0"
]

[project.scripts]
pqcbench = "pqcbench_cli.main:app"

run-kyber = "pqcbench_cli.runners.kyber:app_main"
run-hqc = "pqcbench_cli.runners.hqc:app_main"
run-rsa-oaep = "pqcbench_cli.runners.rsa_oaep:app_main"
run-rsa-pss = "pqcbench_cli.runners.rsa_pss:app_main"
run-dilithium = "pqcbench_cli.runners.dilithium:app_main"
run-falcon = "pqcbench_cli.runners.falcon:app_main"
run-sphincsplus = "pqcbench_cli.runners.sphincsplus:app_main"
run-xmssmt = "pqcbench_cli.runners.xmssmt:app_main"
run-mayo = "pqcbench_cli.runners.mayo:app_main"

[tool.setuptools.packages.find]
where = ["src"]
